UserAndOperationDeserializationTest is moved to dto subpackage
