+++ src/test/java/gov/ca/cwds/idm/UpdateUserIdmResourceTest.java
@@ -45,7 +45,7 @@ import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-public class UserUpdateIdmResourceTest extends BaseIdmResourceTest {
+public class UpdateUserIdmResourceTest extends BaseIdmResourceTest {
 
   @Test
   @WithMockCustomUser
